requirements.txt
Flask==3.1.0
gunicorn==23.0.0
openai==1.66.5
python-dotenv==1.0.1
requests==2.32.3
